export user store into separate module
